+++ src/common/utils/index.ts
@@ -15,6 +15,7 @@ import simpleGit, { FileStatusResult, SimpleGit } from 'simple-git';
 import { CONSTANTS } from '../../config';
 import { MetadataUtils } from '../metadata-utils';
 import { prompts } from './prompts';
+import { encryptFile } from '../cryptoUtils';
 
 let pluginsStdout = null;
 
@@ -56,6 +57,11 @@ export function isGitRepo() {
   return isInsideWorkTree.status === 0;
 }
 
+export async function gitHasLocalUpdates() {
+  const gitStatus = await git().status();
+  return (gitStatus.files.length > 0)
+}
+
 // Install plugin if not present
 export async function checkSfdxPlugin(pluginName: string) {
   // Manage cache of sfdx plugins result
@@ -77,7 +83,7 @@ export async function promptInstanceUrl() {
     message: c.cyanBright('Is the org you need to connect a sandbox or another type of org (dev org, enterprise org...)'),
     choices: [
       { title: 'Sandbox', description: 'The org I want to connect is a sandbox', value: 'https://test.salesforce.com' },
-      { title: 'Other', description: 'The org I want to connect is not a sandbox', value: 'https://login.salesforce.com' }
+      { title: 'Other (Dev org, Enterprise org...)', description: 'The org I want to connect is NOT a sandbox', value: 'https://login.salesforce.com' }
     ],
     initial: 1
   });
@@ -367,7 +373,8 @@ export async function execCommand(
     // if --json, we should not have a crash, so return status 1 + output log
     return {
       status: 1,
-      errorMessage: `[sfdx-hardis][ERROR] Error processing command\n$${e.stdout}\n${e.stderr}`
+      errorMessage: `[sfdx-hardis][ERROR] Error processing command\n$${e.stdout}\n${e.stderr}`,
+      error : e
     };
   }
   // Display output if requested, for better user unrstanding of the logs
@@ -693,7 +700,9 @@ export async function generateSSLCertificate(branchName: string, folder: string,
   process.chdir(prevDir);
   // Copy certificate key in local project
   await fs.ensureDir(folder);
-  await fs.copy(path.join(tmpDir, 'server.key'), path.join(folder, `${branchName}.key`));
+  const targetKeyFile = path.join(folder, `${branchName}.key`);
+  await fs.copy(path.join(tmpDir, 'server.key'), targetKeyFile);
+  const encryptionKey = await encryptFile(targetKeyFile);
   // Copy certificate file in user home project
   const crtFile = path.join(os.homedir(), `${branchName}.crt`);
   await fs.copy(path.join(tmpDir, 'server.crt'), crtFile);
@@ -712,6 +721,7 @@ export async function generateSSLCertificate(branchName: string, folder: string,
   });
   if (confirmResponse.value === true) {
     uxLog(commandThis, c.cyanBright(`You must configure CI variable ${c.green(c.bold(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`))} with value ${c.bold(c.green(consumerKey))}`));
+    uxLog(commandThis, c.cyanBright(`You must configure CI variable ${c.green(c.bold(`SFDX_CLIENT_KEY_${branchName.toUpperCase()}`))} with value ${c.bold(c.green(encryptionKey))}`));
     await prompts({ type: 'confirm', message: c.cyanBright('In GitLab it is in Project -> Settings -> CI/CD -> Variables. Hit ENTER when it is done') });
     // Request info for deployment
     const promptResponses = await prompts([
@@ -780,7 +790,7 @@ export async function generateSSLCertificate(branchName: string, folder: string,
     try {
       const deployRes = await MetadataUtils.deployMetadatas({
         deployDir: tmpDirMd,
-        testlevel: (branchName.includes('production'))?'RunAllTests':'NoTestRun',
+        testlevel: (branchName.includes('production'))?'RunLocalTests':'NoTestRun',
         soap: true
       });
       console.assert(deployRes.status === 0, c.red('[sfdx-hardis] Failed to deploy metadatas'));
@@ -810,6 +820,7 @@ Hit ENTER when you are ready`)
     uxLog(commandThis, c.yellow('Now you can configure the sfdx connected app'));
     uxLog(commandThis, `Follow instructions here: ${c.bold('https://developer.salesforce.com/docs/atlas.en-us.sfdx_dev.meta/sfdx_dev/sfdx_dev_auth_connected_app.htm')}`);
     uxLog(commandThis, `Use ${c.green(crtFile)} as certificate on Connected App configuration page, ${c.bold(`then delete ${crtFile} for security`)}`);
-    uxLog(commandThis, `Then, configure CI variable ${c.green(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`)} with value of ConsumerKey on Connected App configuration page`);
+    uxLog(commandThis, `- configure CI variable ${c.green(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`)} with value of ConsumerKey on Connected App configuration page`);
+    uxLog(commandThis, `- configure CI variable ${c.green(`SFDX_CLIENT_KEY_${branchName.toUpperCase()}`)} with value ${c.green(encryptionKey)} key`);
   }
 }