+++ src/hooks/prerun/auth.ts
@@ -3,6 +3,7 @@ import * as c from 'chalk';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
+import { decryptFile } from '../../common/cryptoUtils';
 import { execCommand, execSfdxJson, getCurrentGitBranch, isCI, promptInstanceUrl, restoreLocalSfdxInfo, uxLog } from '../../common/utils';
 import { checkConfig, getConfig } from '../../config';
 
@@ -62,9 +63,9 @@ async function authOrg(orgAlias: string, options: any) {
     if (!options.checkAuth) {
         // Check if we are already authenticated
         let orgDisplayCommand = 'sfdx force:org:display';
-        if (orgAlias !== 'MY_ORG') {
+        if (orgAlias !== 'MY_ORG' && (isCI||isDevHub)) {
             orgDisplayCommand += ' --targetusername ' + orgAlias;
-        }
+        } 
         const orgInfoResult = await execSfdxJson(orgDisplayCommand, this, { fail: false, output: false, debug: options.debug });
         if (
             orgInfoResult.result &&
@@ -72,6 +73,7 @@ async function authOrg(orgAlias: string, options: any) {
                 orgInfoResult.result.connectedStatus.includes('Connected')) ||
                 (options.scratch && orgInfoResult.result.connectedStatus.includes('Unknown')) ||
                 (orgInfoResult.result.alias === orgAlias && orgInfoResult.result.id != null) ||
+                (orgInfoResult.result.username === orgAlias && orgInfoResult.result.id != null) ||
                 (isDevHub && orgInfoResult.result.id != null))
         ) {
             // Set as default username or devhubusername
@@ -122,7 +124,7 @@ async function authOrg(orgAlias: string, options: any) {
                         : 'https://login.salesforce.com';
         // Get JWT items clientId and certificate key
         const sfdxClientId = await getSfdxClientId(orgAlias, config);
-        const crtKeyfile = await getCertificateKeyFile(orgAlias);
+        const crtKeyfile = await getCertificateKeyFile(orgAlias,config);
         const usernameArg =
             isDevHub
                 ? '--setdefaultdevhubusername'
@@ -138,6 +140,7 @@ async function authOrg(orgAlias: string, options: any) {
                 ` --setalias ${orgAlias}` +
                 ` --instanceurl ${instanceUrl}`;
             const jwtAuthRes = await execSfdxJson(loginCommand, this, { fail: false });
+            // await fs.remove(crtKeyfile); // Delete private key file from temp folder TODO: move to postrun hook
             logged = jwtAuthRes.status === 0;
             if (!logged) {
                 console.error(c.red(`[sfdx-hardis][ERROR] JWT login error: \n${JSON.stringify(jwtAuthRes)}`));
@@ -221,8 +224,37 @@ async function getSfdxClientId(orgAlias: string, config: any) {
     return null;
 }
 
+// Get clientId for SFDX connected app
+async function getKey(orgAlias: string, config: any) {
+    // Try to find in global variables
+    const sfdxClientKeyVarName = `SFDX_CLIENT_KEY_${orgAlias}`;
+    if (process.env[sfdxClientKeyVarName]) {
+        return process.env[sfdxClientKeyVarName];
+    }
+    const sfdxClientKeyVarNameUpper = sfdxClientKeyVarName.toUpperCase();
+    if (process.env[sfdxClientKeyVarNameUpper]) {
+        return process.env[sfdxClientKeyVarNameUpper];
+    }
+    if (process.env.SFDX_CLIENT_KEY) {
+        console.warn(
+            c.yellow(`[sfdx-hardis] If you use CI on multiple branches & orgs, you should better define CI variable ${c.bold(sfdxClientKeyVarNameUpper)} than SFDX_CLIENT_KEY`)
+        );
+        return process.env.SFDX_CLIENT_KEY;
+    }
+    // Try to find in config files ONLY IN LOCAL MODE (in CI, it's supposed to be a CI variable)
+    if (!isCI && config.devHubSfdxClientKey) {
+        return config.devHubSfdxClientKey;
+    }
+    if (isCI) {
+        console.error(
+            c.red(`[sfdx-hardis] You must set env variable ${c.bold(sfdxClientKeyVarNameUpper)} with the value of SSH private key encryption key`)
+        );
+    }
+    return null;
+}
+
 // Try to find certificate key file for sfdx connected app in different locations
-async function getCertificateKeyFile(orgAlias: string) {
+async function getCertificateKeyFile(orgAlias: string, config: any) {
     const filesToTry = [
         `./config/branches/.jwt/${orgAlias}.key`,
         `./config/.jwt/${orgAlias}.key`,
@@ -232,7 +264,11 @@ async function getCertificateKeyFile(orgAlias: string) {
     ];
     for (const file of filesToTry) {
         if (fs.existsSync(file)) {
-            return file;
+            // Decrypt SSH private key and write a temporary file
+            const sshKey = await getKey(orgAlias,config); 
+            const tmpSshKeyFile = path.join(os.tmpdir(), `${orgAlias}.key`); 
+            await decryptFile(file,tmpSshKeyFile,sshKey);
+            return tmpSshKeyFile; 
         }
     }
     if (isCI) {